Factor out s3 upload function
