+++ app/workers/generate_global_export_csv_worker.rb
@@ -1,3 +1,5 @@
+require "s3_file_uploader"
+
 class GenerateGlobalExportCsvWorker
   include Sidekiq::Worker
 
@@ -8,24 +10,8 @@ class GenerateGlobalExportCsvWorker
       export_params["exclude_spam"],
     ).call
 
-    s3_file = self.class.save_file_to_s3(filename, contents)
+    s3_file = S3FileUploader.save_file_to_s3(filename, contents)
 
     GlobalExportNotification.notification_email(export_params["notification_email"], s3_file.key).deliver_now
-  end
-
-  def self.save_file_to_s3(filename, csv)
-    connection = Fog::Storage.new(
-      provider: "AWS",
-      region: ENV["AWS_REGION"],
-      aws_access_key_id: ENV["AWS_ACCESS_KEY_ID"],
-      aws_secret_access_key: ENV["AWS_SECRET_ACCESS_KEY"],
-    )
-
-    directory = connection.directories.get(ENV["AWS_S3_BUCKET_NAME"])
-
-    directory.files.create(
-      key: filename,
-      body: csv,
-    )
   end
 end
